+++ runtime/src/f2j/unbox/ClosureIntInt.java
@@ -2,8 +2,8 @@ package f2j.unbox;
 
 public abstract class ClosureIntInt implements Closure
 {
-  public int x;
-  public int out;
+  public long x;
+  public long out;
   public abstract void apply ()
   ;
   public ClosureIntInt clone () {
